rename: board -> visitor board
