+++ config.toml
@@ -9,7 +9,7 @@ log_to_stdout = true # additional logging to STDOUT
 aero_host = "aerospike" # hostname of aerospike server
 aero_port = 3000 # aerospike server port number
 aero_namespace = "serj-tubin-com" # aerospike namespace value (used in aerospike server)
-aero_messages_set = "messages" # aerospike set name for board messages (used in aerospike server)
+aero_messages_set = "messages" # aerospike set name for visitor_board messages (used in aerospike server)
 # NETLOG BACKUP
 netlog_unix_socket_addr_dir  = "/var/tmp/serj-service"
 netlog_unix_socket_file_name = "netlog-backup.sock"
@@ -37,7 +37,7 @@ log_to_stdout = true # additional logging to STDOUT
 aero_host = "172.28.128.3" # hostname of aerospike server
 aero_port = 3000 # aerospike server port number
 aero_namespace = "test" # aerospike namespace value (used in aerospike server)
-aero_messages_set = "messages" # aerospike set name for board messages (used in aerospike server)
+aero_messages_set = "messages" # aerospike set name for visitor_board messages (used in aerospike server)
 # NETLOG BACKUP
 netlog_unix_socket_addr_dir  = "/var/tmp/serj-service"
 netlog_unix_socket_file_name = "netlog-backup.sock"
@@ -65,7 +65,7 @@ log_to_stdout = false # additional logging to STDOUT
 aero_host = "localhost" # hostname of aerospike server
 aero_port = 3000 # aerospike server port number
 aero_namespace = "serj-tubin-com" # aerospike namespace value (used in aerospike server)
-aero_messages_set = "messages" # aerospike set name for board messages (used in aerospike server)
+aero_messages_set = "messages" # aerospike set name for visitor_board messages (used in aerospike server)
 # NETLOG BACKUP
 netlog_unix_socket_addr_dir  = "/var/tmp/serj-service" # TODO: change to use /var/run
 netlog_unix_socket_file_name = "netlog-backup.sock"
