+++ internal/visitor_board/aerospike/board_aero_test_client.go
