+++ internal/visitor_board/cache.go
@@ -1,4 +1,4 @@
-package board
+package visitor_board
 
 import (
 	"fmt"
