+++ internal/visitor_board/handler_test.go
@@ -1,4 +1,4 @@
-package board
+package visitor_board
 
 import (
 	"bytes"
