+++ internal/visitor_board/test_cache.go
@@ -1,4 +1,4 @@
-package board
+package visitor_board
 
 import (
 	"sync"
